Rework Manual Purchases to act as a Virtual Terminal.
The process for making a manual purchase now resembles the order process.
Products are selected and added to a cart, shipping or billing address are
selected, shipping method is chosen.

All of the REST enabled payment gateways can be used to purchase the cart. New
credit card information can be entered from the page or a payment token can be
used. The manual purchases gateway can also be used.

Todo:
- Allow adding a new address.
- Allow adjusting the pre-tax total.
- Allow adding coupons.
- (maybe) Allow adding fees.